+++ .php-cs-fixer.dist.php
@@ -4,7 +4,7 @@ $finder = PhpCsFixer\Finder::create()
     ->exclude('public')
     ->in(__DIR__);
 
-return PhpCsFixer\Config::create()
+return (new PhpCsFixer\Config)
     ->setRiskyAllowed(true)
     ->setRules([
         '@PSR2' => true,