+++ changelog.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Remove double dots from email content rows start (fix #13)
+
 ## [1.0.4] - 2021-04-02
 
 ### Added